Trademesh/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cssselect==1.3.0
cymem==2.0.11
discord.py==2.5.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
holidays==0.70
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.2
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
multitasking==0.0.11
murmurhash==1.0.12
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
preshed==3.0.9
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
ta==0.11.0
textblob==0.19.0
thinc==8.3.6
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.19.0
yfinance==0.2.55
